Code generator refactors (#1492)
* Remove duplicate code, refactor similar functions to combine
* Move some classes into separate files
* Components MAC code for all remaining data types
* Update MAC code formatting for readability
* New asmCaps for fma_f64 and dot4
